Split columns and parameters
